ingest/requirements.txt
pandas
python-dotenv
elasticsearch7